Remove ifel and droplevels.
ifel is not necessary since with just a tiny
bit of extra care we can use tidyverse's
if_else to to the same thing, and arguably
since the method is more standard and also
documented, also clearer.

Since we are using tidyverse by and large,
changing the base droplevels seems largely
unnecessary. Instead will fashion a wraper
(possibly sharing the name) arounf fct_drop
which will do a similar thing.